requirements/base.txt
Django==2.0.1
psycopg2-binary==2.8.5
python-decouple==3.1
pytz==2019.3
uWSGI==2.0.15
django-widget-tweaks==1.4.2
django-wysiwyg==0.8.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-taggit==1.2.0
django-bootstrap4==1.1.1